backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
supabase>=2.0.2
langchain>=0.0.350
langchain-google-genai>=0.0.5
google-generativeai>=0.3.2
PyPDF2>=3.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.0
httpx>=0.25.2
numpy>=1.26.0
pandas>=2.2.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
alembic>=1.12.1
python-dateutil>=2.8.2
